requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-activity-stream==0.8.0
django-appconf==1.0.4
django-categories==1.7.1
django-compressor==2.4
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-filter==2.2.0
django-haystack==3.0b1
django-js-asset==1.2.2
django-lazysignup==2.0.0
django-libsass==0.8
django-mptt==0.10.0
django-seed==0.2.2
djangorestframework==3.11.0
Faker==4.1.0
idna==2.9
libsass==0.20.0
Markdown==3.2.2
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.9
Whoosh==2.7.4
